+++ src/main/kotlin/com/github/jensim/megamanipulator/actions/vcs/RepoWrapper.kt
@@ -11,6 +11,7 @@ sealed class RepoWrapper {
     abstract fun getProject(): String
     abstract fun getRepo(): String
     abstract fun getCloneUrl(): String?
+    abstract fun getDefaultBranch(): String?
     fun asPathString() = "${getSearchHost()}/${getCodeHost()}/${getProject()}/${getRepo()}"
 }
 
@@ -19,12 +20,14 @@ data class BitBucketRepoWrapping(
     private val searchHost: String,
     private val codeHost: String,
     val repo: BitBucketRepo,
+    private val defaultBranch: String?,
 ) : RepoWrapper() {
     override fun getSearchHost(): String = searchHost
     override fun getCodeHost(): String = codeHost
     override fun getProject(): String = repo.project!!.key
     override fun getRepo(): String = repo.slug
     override fun getCloneUrl(): String? = repo.links?.clone?.first { it.name == "ssh" }?.href
+    override fun getDefaultBranch(): String? = defaultBranch
 }
 
 @Serializable
@@ -38,4 +41,5 @@ data class GithubComRepoWrapping(
     override fun getProject(): String = repo.owner.login
     override fun getRepo(): String = repo.name
     override fun getCloneUrl(): String = repo.ssh_url
+    override fun getDefaultBranch(): String = repo.default_branch
 }